platform: prometheus: release cx6_reset at boot
There was an issue observed that after a reboot command we would no
longer see devices downstream from the cx6.  To fix the issue there are
two pieces:

1. Initial pinmux for GPIO AA.00 is updated to drive low.  That happens
during the early bootloader.  That will put the CX6 into a consistent
state on both power-up and reboot.

2. In this patch we register a fixed regulator with the PCIe controller
corresponding to the CX6.  This regulator doesn't physically exist. The
intent is that Linux turns on the PCIe power to enable this "regulator"
which will have the effect of releasing reset for the cx6 since the
associated GPIO enable pin for this regulator connects to the CX6 RST.

Prior to this patch GPIO AA.00 was in a high impedance state all the
time and so the reset of the CX6 was purely hardware controlled.  This
patch configures the reset pin as open drain such that we drive it low
initially, but set it high impedance when we are ready to release the
CX6 from reset.  The reset circuitry will bring the signal high.  With
these fixes in place, the Tegra and the CX6 will always be in a
consistent state for both cold boot and reboot.

Bug 3464723
Bug 3400304

Change-Id: I37e20672346e253e0d8b256e0df12ad9ab079df4
Signed-off-by: Brad Griffis <bgriffis@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t23x/prometheus-dts/+/2639609
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Manikanta Maddireddy <mmaddireddy@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit